+++ utils/navigation.js
@@ -6,6 +6,17 @@ function buildFileUrl(parentHash, path) {
   return `/content/${parentHash}/${path}`;
 }
 
+function buildObjectUrl(parentHash, { path, type }) {
+  switch (type) {
+    case 'tree':
+      return buildFolderUrl(parentHash, path);
+    case 'blob':
+      return buildFileUrl(parentHash, path);
+    default:
+      return '#';
+  }
+}
+
 function buildBreadcrumbs(hash, path) {
   const bc = [
     {
@@ -48,5 +59,6 @@ function buildBreadcrumbs(hash, path) {
 module.exports = {
   buildFolderUrl,
   buildFileUrl,
+  buildObjectUrl,
   buildBreadcrumbs
 };
